add file reference query to value object
